% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emrsim-package.R
\docType{package}
\name{emrsim-package}
\alias{emrsim-package}
\alias{emrsim}
\title{Simulate EMR Data}
\description{
Package form simulating equipment maintenance and repair data.
}
\author{
Andreas Nordland (Maintainer) <andreas.nordland@gmail.com>
}
\keyword{package}
